+++ .gitignore
@@ -37,7 +37,7 @@ bower_components
 # Compiled binary addons (https://nodejs.org/api/addons.html)
 build/Release
 
-# Dependency directories
+# Index directories
 node_modules/
 jspm_packages/
